refactor: organização das views em pastas
