+++ src/views/Dashboard/Dashbord.vue
@@ -1,4 +1,4 @@
 <template>
-  <h1>Admin Home </h1>
+  <h1>Home </h1>
   <img src="https://s2.glbimg.com/dWcTfyFzZxj6dL-I-RK7NZTXebs=/475x315/top/i.glbimg.com/og/ig/infoglobo/f/original/2016/02/17/homer.jpg" width="100px" />
 </template>
