+++ src/views/Login/Login.vue
@@ -58,10 +58,10 @@
 </template>
 
 <script>
-import api from '../services/api'
+import api from '../../services/api'
 import * as yup from 'yup'
-import { captureErrorYup } from '../utils/captureErrorYup.js'
-import AuthenticationService from '../services/AuthenticationService'
+import { captureErrorYup } from '../../utils/captureErrorYup'
+import AuthenticationService from '../../services/AuthenticationService'
 import router from '@/router'
 
 export default {
